+++ JsonToolsNppPlugin/Properties/AssemblyInfo.cs
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("4.14.0.14")]
-[assembly: AssemblyFileVersion("4.14.0.14")]
+[assembly: AssemblyVersion("4.14.0.15")]
+[assembly: AssemblyFileVersion("4.14.0.15")]